requirements.txt
Flask==2.1.2
pdfplumber==0.5.28
python-dotenv==0.19.2
google-generativeai==0.1.4
fpdf==1.7.2
werkzeug==2.1.2
